o Added debug logging to better track issues with passivation & activation  o Using UID for id for better uniqness  o Extending ServiceMBeanSupport for better state management  o Fixed up the rare case of a non-Exception/non-Error being thrown  o Removed unused field related fluff (private stuff not being used, looks    like SessionInputStream & SessionOutputStream have superceeded it).

+++ src/main/org/jboss/ejb/plugins/StatefulSessionFilePersistenceManager.java
@@ -7,23 +7,22 @@
 
 package org.jboss.ejb.plugins;
 
+import java.io.IOException;
 import java.io.File;
 import java.io.InputStream;
 import java.io.OutputStream;
 import java.io.ObjectInputStream;
 import java.io.ObjectOutputStream;
+import java.io.FileInputStream;
 import java.io.FileOutputStream;
 import java.io.BufferedInputStream;
 import java.io.BufferedOutputStream;
-import java.io.FileInputStream;
-import java.io.IOException;
+
 import java.lang.reflect.Method;
-import java.lang.reflect.Field;
-import java.lang.reflect.Modifier;
 import java.lang.reflect.InvocationTargetException;
+
 import java.rmi.RemoteException;
 import java.rmi.ServerException;
-import java.util.ArrayList;
 
 import javax.ejb.Handle;
 import javax.ejb.EJBObject;
@@ -35,57 +34,68 @@ import org.jboss.ejb.Container;
 import org.jboss.ejb.StatefulSessionContainer;
 import org.jboss.ejb.StatefulSessionPersistenceManager;
 import org.jboss.ejb.StatefulSessionEnterpriseContext;
-import org.jboss.logging.Logger;
 
 import org.jboss.system.server.ServerConfigLocator;
+import org.jboss.system.ServiceMBeanSupport;
 
 import org.jboss.util.jmx.MBeanServerLocator;
 
+import org.jboss.util.id.UID;
+
 /**
  * This class is one of the passivating plugins for JBoss.
+ *
+ * <p>
  * It is fairly simple and can work from the file system from wich JBoss is operating.
  * Passivates beans into the system data directory under the 'sessions' sub-directory.
  *
+ * @jmx:mbean extends="org.jboss.system.ServiceMBean"
+ * 
+ * @version <tt>$Revision: 1.37 $</tt>
  * @author <a href="mailto:rickard.oberg@telkel.com">Rickard Öberg</a>
  * @author <a href="mailto:marc.fleury@telkel.com">Marc Fleury</a>
  * @author <a href="mailto:sebastien.alborini@m4x.org">Sebastien Alborini</a>
- * @version $Revision: 1.36 $
+ * @author <a href="mailto:jason@planet57.com">Jason Dillon</a>
  */
 public class StatefulSessionFilePersistenceManager
-   implements StatefulSessionPersistenceManager
+   extends ServiceMBeanSupport
+   implements StatefulSessionPersistenceManager, StatefulSessionFilePersistenceManagerMBean
 {
-   private static Logger log = Logger.getLogger(StatefulSessionFilePersistenceManager.class);
-   
-   private StatefulSessionContainer con;
-    
-   private Method ejbActivate;
-   private Method ejbPassivate;
-   private Method ejbRemove;
+   /** A refernece to the {@link SessionBean#ejbActivate} method. */
+   private static final Method EJB_ACTIVATE_METHOD;
 
+   /** A refernece to the {@link SessionBean#ejbPassivate} method. */
+   private static final Method EJB_PASSIVATE_METHOD;
+   
+   /** A refernece to the {@link SessionBean#ejbRemove} method. */
+   private static final Method EJB_REMOVE_METHOD;
+
+   /** Setup method references. */
+   static {
+      try {
+         EJB_ACTIVATE_METHOD = SessionBean.class.getMethod("ejbActivate", new Class[0]);
+         EJB_PASSIVATE_METHOD = SessionBean.class.getMethod("ejbPassivate", new Class[0]);
+         EJB_REMOVE_METHOD = SessionBean.class.getMethod("ejbRemove", new Class[0]);
+      }
+      catch (Exception e) {
+         throw new org.jboss.util.NestedError("System failure; Missing required interface classes", e);
+      }
+   }
+
+   /** Our container. */
+   private StatefulSessionContainer con;
+
+   /** The base directory where sessions state files are stored for our container. */
    private File dataDir;
-    
-   private ArrayList fields;
-    
-   private static long id = System.currentTimeMillis();
     
    public void setContainer(Container c)
    {
       con = (StatefulSessionContainer)c;
    }
-    
-   public void create()
-      throws Exception
+
+   protected void createService() throws Exception
    {
       boolean debug = log.isDebugEnabled();
-
-      // FIXME: these never change, make them static and loaded on class init
-
-      // Find methods
-      ejbActivate = SessionBean.class.getMethod("ejbActivate", new Class[0]);
-
-      ejbPassivate = SessionBean.class.getMethod("ejbPassivate", new Class[0]);
-
-      ejbRemove = SessionBean.class.getMethod("ejbRemove", new Class[0]);
 
       // Initialize the dataStore
       String ejbName = con.getBeanMetaData().getEjbName();
@@ -96,62 +106,53 @@ public class StatefulSessionFilePersistenceManager
       // Gte a handle on the sysdata/sessions/ejbname directory
       dataDir = new File(new File(systemDataDir, "sessions"), ejbName);
       if (debug) {
-         log.debug("Storing sessions for " + ejbName + " in: " + dataDir);
+         log.debug("Storing sessions for '" + ejbName + "' in: " + dataDir);
       }
 
-      dataDir.mkdirs();
+      // if the directory does not exist then try to create it
+      if (!dataDir.exists()) {
+         if (!dataDir.mkdirs()) {
+            throw new IOException("Failed to create directory: " + dataDir);
+         }
+      }
+      
+      // make sure we have a directory
+      if (!dataDir.isDirectory()) {
+         throw new IOException("File exists where directory expected: " + dataDir);
+      }
 
-      // Clear dir of old files
+      // make sure we can read and write to it
+      if (!dataDir.canWrite() || !dataDir.canRead()) {
+         throw new IOException("Directory must be readable and writable: " + dataDir);
+      }
+      
+      // Purge state session state files
       File[] sessions = dataDir.listFiles();
       for (int i = 0; i < sessions.length; i++)
       {
          if (! sessions[i].delete()) {
             log.warn("Failed to delete session state file: " + sessions[i]);
          }
+         else {
+            log.debug("Removed stale session state: " + sessions[i]);
+         }
       }
-      
-      if (debug) {
-         log.debug(sessions.length + " old sessions removed");
-      }
+   }
 
-      // Get fields of class
-      Class beanClass = con.getBeanClass();
-      fields = new ArrayList();
-      
-      while (!beanClass.equals(Object.class))
-      {
-         Field[] f = beanClass.getDeclaredFields();
-            
-         // Skip transient fields
-         for (int i = 0; i < f.length; i++)
-            if (!Modifier.isTransient(f[i].getModifiers()))
-            {
-               fields.add(f[i]);
-            }
-            
-         beanClass = beanClass.getSuperclass();
-      }
-   }
-    
-   public void start()
-      throws Exception
+   /**
+    * Make a file for the given instance id.
+    */
+   protected File getFile(final Object key)
    {
+      return new File(dataDir, key + ".ser");
    }
-    
-   public void stop()
-   {
-   }
-    
-   public void destroy()
-   {
-   }
-    
+
    public void createSession(Method m, Object[] args, StatefulSessionEnterpriseContext ctx)
       throws Exception
    {
-        
       // Set id
-      ctx.setId(nextId());
+      Object id = new UID();
+      ctx.setId(id);
         
       // Get methods
       try
@@ -161,15 +162,16 @@ public class StatefulSessionFilePersistenceManager
             
          // Call ejbCreate<METHOD>
          createMethod.invoke(ctx.getInstance(), args);
-        
-      } catch (IllegalAccessException e)
+      }
+      catch (IllegalAccessException e)
       {
          // Clear id
          ctx.setId(null);
             
          // Throw this as a bean exception...(?)
          throw new EJBException(e);
-      } catch (InvocationTargetException ite)
+      }
+      catch (InvocationTargetException ite)
       {
          // Clear id
          ctx.setId(null);
@@ -190,9 +192,12 @@ public class StatefulSessionFilePersistenceManager
             // Remote, Create, or custom app. exception
             throw (Exception)e;
          }
-         else
+         else if (e instanceof Error)
          {
             throw (Error)e;
+         }
+         else {
+            throw new org.jboss.util.UnexpectedThrowable(e);
          }
       }
         
@@ -201,48 +206,67 @@ public class StatefulSessionFilePersistenceManager
         
       // Create EJBObject
       if (con.getProxyFactory() != null)
-         ctx.setEJBObject((EJBObject)con.getProxyFactory().getStatefulSessionEJBObject(ctx.getId()));
+         ctx.setEJBObject((EJBObject)con.getProxyFactory().getStatefulSessionEJBObject(id));
+      
       // Create EJBLocalObject
       if (con.getLocalHomeClass() != null)
-         ctx.setEJBLocalObject(con.getLocalProxyFactory().getStatefulSessionEJBLocalObject(ctx.getId()));
+         ctx.setEJBLocalObject(con.getLocalProxyFactory().getStatefulSessionEJBLocalObject(id));
    }
-    
+
    public void activateSession(StatefulSessionEnterpriseContext ctx)
       throws RemoteException
    {
       try
       {
-         if (log.isDebugEnabled()) {
+         boolean debug = log.isDebugEnabled();
+         if (debug) {
             log.debug("Attempting to activate; ctx=" + ctx);
          }
-         
-         ObjectInputStream in;
-            
+
+         Object id = ctx.getId();
+
          // Load state
-         in = new SessionObjectInputStream
-            (ctx, new BufferedInputStream
-             (new FileInputStream(getFile(ctx.getId()))));
+         File file = getFile(id);
+         if (debug) {
+            log.debug("Reading object state from: " + file);
+         }
+
+         ObjectInputStream in =
+            new SessionObjectInputStream(ctx, new BufferedInputStream(new FileInputStream(file)));
+         
+         try {
+            Object obj = in.readObject();
+            log.debug("Object state: " + obj);
             
-         ctx.setInstance(in.readObject());
+            ctx.setInstance(obj);
+         }
+         finally {
+            in.close();
+         }
             
-         in.close();
-            
-         removePassivated(ctx.getId());
+         removePassivated(id);
             
          // Call bean
-         ejbActivate.invoke(ctx.getInstance(), new Object[0]);
-        
-      } catch (ClassNotFoundException e)
+         EJB_ACTIVATE_METHOD.invoke(ctx.getInstance(), new Object[0]);
+
+         if (debug) {
+            log.info("Activation complete; ctx=" + ctx);
+         }
+      }
+      catch (ClassNotFoundException e)
       {
          throw new ServerException("Could not activate", e);
-      } catch (IOException e)
+      }
+      catch (IOException e)
       {
          throw new ServerException("Could not activate", e);
-      } catch (IllegalAccessException e)
+      }
+      catch (IllegalAccessException e)
       {
          // Throw this as a bean exception...(?)
          throw new EJBException(e);
-      } catch (InvocationTargetException ite)
+      }
+      catch (InvocationTargetException ite)
       {
          Throwable e = ite.getTargetException();
          if (e instanceof EJBException)
@@ -260,9 +284,12 @@ public class StatefulSessionFilePersistenceManager
             // Remote, Create, or custom app. exception
             throw (RemoteException)e;
          }
-         else
+         else if (e instanceof Error)
          {
             throw (Error)e;
+         }
+         else {
+            throw new org.jboss.util.UnexpectedThrowable(e);
          }
       }
    }
@@ -272,31 +299,50 @@ public class StatefulSessionFilePersistenceManager
    {
       try
       {
-         if (log.isDebugEnabled()) {
+         boolean debug = log.isDebugEnabled();
+         if (debug) {
             log.debug("Attempting to passivate; ctx=" + ctx);
          }
          
          // Call bean
-         ejbPassivate.invoke(ctx.getInstance(), new Object[0]);
+         EJB_PASSIVATE_METHOD.invoke(ctx.getInstance(), new Object[0]);
             
          // Store state
+
+         File file = getFile(ctx.getId());
+         if (debug) {
+            log.debug("Saving object state to: " + file);
+         }
+         
          ObjectOutputStream out =
-            new SessionObjectOutputStream
-            (new BufferedOutputStream
-             (new FileOutputStream(getFile(ctx.getId()))));
+            new SessionObjectOutputStream(new BufferedOutputStream(new FileOutputStream(file)));
+
+         try {
+            Object obj = ctx.getInstance();
+            if (debug) {
+               log.debug("Writing object state: " + obj);
+            }
             
-         out.writeObject(ctx.getInstance());
-            
-         out.close();
-        
-      } catch (IOException e)
+            out.writeObject(obj);
+         }
+         finally {
+            out.close();
+         }
+
+         if (debug) {
+            log.debug("Passivation complete; ctx=" + ctx);
+         }
+      }
+      catch (IOException e)
       {
          throw new ServerException("Could not passivate", e);
-      }catch (IllegalAccessException e)
+      }
+      catch (IllegalAccessException e)
       {
          // Throw this as a bean exception...(?)
          throw new EJBException(e);
-      } catch (InvocationTargetException ite)
+      }
+      catch (InvocationTargetException ite)
       {
          Throwable e = ite.getTargetException();
          if (e instanceof EJBException)
@@ -314,9 +360,12 @@ public class StatefulSessionFilePersistenceManager
             // Remote, Create, or custom app. exception
             throw (RemoteException)e;
          }
-         else
+         else if (e instanceof Error)
          {
             throw (Error)e;
+         }
+         else {
+            throw new org.jboss.util.UnexpectedThrowable(e);
          }
       }
    }
@@ -324,16 +373,26 @@ public class StatefulSessionFilePersistenceManager
    public void removeSession(StatefulSessionEnterpriseContext ctx)
       throws RemoteException, RemoveException
    {
-      // Call bean
       try
       {
-         ejbRemove.invoke(ctx.getInstance(), new Object[0]);
-        
-      } catch (IllegalAccessException e)
+         boolean debug = log.isDebugEnabled();
+         if (debug) {
+            log.debug("Attempting to remove; ctx=" + ctx);
+         }
+
+         // Call bean
+         EJB_REMOVE_METHOD.invoke(ctx.getInstance(), new Object[0]);
+
+         if (debug) {
+            log.debug("Removal complete; ctx=" + ctx);
+         }
+      }
+      catch (IllegalAccessException e)
       {
          // Throw this as a bean exception...(?)
          throw new EJBException(e);
-      } catch (InvocationTargetException ite)
+      }
+      catch (InvocationTargetException ite)
       {
          Throwable e = ite.getTargetException();
          if (e instanceof EJBException)
@@ -350,30 +409,28 @@ public class StatefulSessionFilePersistenceManager
          {
             throw (RemoveException)e;
          }
-         else if (e instanceof RemoteException)
-         {
-            throw (RemoteException)e;
-         }
-         else
+         else if (e instanceof Error)
          {
             throw (Error)e;
+         }
+         else {
+            throw new org.jboss.util.UnexpectedThrowable(e);
          }
       }
    }
     
    public void removePassivated(Object key)
    {
+      boolean debug = log.isDebugEnabled();
+      
       // OK also if the file does not exists
-      getFile(key).delete();
-   }
-    
-   protected Long nextId()
-   {
-      return new Long(id++);
-   }
-    
-   protected File getFile(Object key)
-   {
-      return new File(dataDir, key + ".ser");
+      File file = getFile(key);
+      if (debug) {
+         log.debug("Removing passivated state file: " + file);
+      }
+      
+      if (file.exists() && !file.delete()) {
+         log.warn("Failed to delete passivated state file: " + file);
+      }
    }
 }
